Automated g4 rollback of changelist 669103619.
*** Reason for rollback ***

Broke TAP tests the LegoML project

*** Original change description ***

Mark message types as requiring the go/jspb object format methods.

This CL marks types that use the JSPB object format accessors (see go/jspb-api-gotchas#objects) so that we do not remove them in the future. This is a backwards compatibility option and will represent no immediate change.

The implementations of these methods are large and expensive for the JSCompiler to process; so we intend to limit their generation to improve compilation performance. See go/lsc-constrain-jspb-object-format-us...

***

PiperOrigin-RevId: 670640538